Removed UnassignedCategory constant from SharedConstantsAndTypes module and replaced references to it with Budget.UnassignedPayeeType
Also changed the name of Budget.PayeeCategory() to Budget.FetchOrAddPayeeCategory() and modified the code to make it more DRY.

Before this change Budget.PayeeCategory() and Budget.GetAssignedCategory() did nearly the same thing (with PayeeCategory() also adding the payee to a category if it came up as not found), and the code was repeated in both functions. Now Budget.FetchOrAddPayeeCategory() calls Budget.GetAssignedCategory() so the code is only defined once.